Split ai to datamuse and gpt3
