+++ functions/src/datamuse/datamuse.ts
@@ -1,30 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import * as functions from 'firebase-functions';
-import { Configuration, OpenAIApi } from 'openai';
-
 import * as https from 'https';
 
 /* eslint-disable @typescript-eslint/no-var-requires */
 const nlp = require('compromise/two');
 const cowsay = require('cowsay');
-
-const openai_key = process.env.OPENAI_SECRET;
-
-const configuration = new Configuration({
-  apiKey: openai_key,
-});
-const openai = new OpenAIApi(configuration);
-
-async function genIdeaOAI(text: string) {
-  const response = await openai.createCompletion({
-    model: 'text-davinci-002',
-    prompt: `Single word or phrase related to ${text}`,
-    temperature: 0.5,
-    max_tokens: 6,
-    n: 3,
-  });
-  return formatIdeaOAI(response.data);
-}
 
 // This method attempts to extract the greatest possible value keywords from the text
 function extractKeywords(text: string) {
@@ -79,31 +59,19 @@ function formatIdeaDM(idea: any) {
   return ideas;
 }
 
-function formatIdeaOAI(idea: any) {
-  // builds an array of strings from the OpenAI response
-  // trim the whitespace and lowercase all the strings
-  let ideas = idea.choices.map((item: any) => item.text.trim().toLowerCase());
-  // remove duplicates
-  ideas = ideas.filter((item: any, pos: any) => ideas.indexOf(item) === pos);
-  return ideas;
-}
-
 /* eslint-disable no-console */
-const ai = functions
+const datamuse = functions
   .runWith({ secrets: ['OPENAI_SECRET'] })
   .region('us-west2')
   .https.onCall(async (data) => {
-    const result = await genIdeaOAI(data.data);
     const resultDM = await genIdeaDM(data.data);
     console.log(cowsay.say({ text: 'Success!' }));
     console.log(cowsay.say(`Datamuse Thinks: ${{ text: resultDM[0] }}`));
-    console.log(cowsay.say(`OpenAI Thinks: ${{ text: result[0] }}`));
     return {
       idea: {
-        openai: result,
         datamuse: resultDM,
       },
     };
   });
 
-export default ai;
+export default datamuse;
