+++ functions/src/index.ts
@@ -7,3 +7,5 @@ initializeApp();
 
 // eslint-disable-next-line import/prefer-default-export
 export { default as ai } from './ai/ai.js';
+export { default as datamuse } from './datamuse/datamuse.js';
+export { default as gpt3 } from './gpt3/gpt3.js';
